+++ waspc/cli/src/Wasp/Cli/Command.hs
@@ -29,18 +29,23 @@ import Control.Monad.Error.Class (MonadError)
 import Control.Monad.Except (ExceptT, runExceptT)
 import Control.Monad.IO.Class (MonadIO)
 import Control.Monad.State.Strict (StateT, evalStateT, gets, modify)
+import Control.Monad.Trans.Resource (MonadResource, ResourceT, runResourceT)
 import Data.Data (Typeable, cast)
 import Data.Maybe (mapMaybe)
 import System.Exit (exitFailure)
 import Wasp.Cli.Message (cliSendMessage)
 import qualified Wasp.Message as Msg
 
-newtype Command a = Command {_runCommand :: StateT [Requirement] (ExceptT CommandError IO) a}
-  deriving (Functor, Applicative, Monad, MonadIO, MonadError CommandError)
+newtype Command a = Command
+  { _runCommand ::
+      StateT [Requirement] (ExceptT CommandError (ResourceT IO)) a
+  }
+  deriving (Functor, Applicative, Monad, MonadIO, MonadError CommandError, MonadResource)
 
 runCommand :: Command a -> IO ()
 runCommand cmd = do
-  runExceptT (flip evalStateT [] $ _runCommand cmd) >>= \case
+  result <- runResourceT $ runExceptT $ (`evalStateT` []) $ _runCommand cmd
+  case result of
     Left cmdError -> do
       cliSendMessage $ Msg.Failure (_errorTitle cmdError) (_errorMsg cmdError)
       exitFailure